pyproject.toml
[project]
name = "fivetran-custom-connector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.2"
dependencies = [
    "fivetran-connector-sdk>=1.3.0",
    "snowflake-connector-python>=3.14.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "duckdb>=1.2.2",
    "isort>=6.0.1",
    "jupyterlab>=4.4.0",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.4",
]
